Add Src relocation model + convert action to a fake instead of mock.
Fix up flow diagram to only remove the node at the very end.
Add minimum testing for src.

Already we can see an issue with the flow diagram:
In the dst section, if Candidate Info RPC is received before we
consensus ExpectCandidate, we will discard it, if we are one of the
good node, bad node can reject the node by votting purge since we will
never receive CandidateInfo again.
  => We need to eliminate these race conditions