fix/test: remove side-effect from lose_proxy_connection
By the way it was constructed, the test `lose_proxy_connection` was
enforcing a specific number of network or events.
This coupling was unrelated to the core logic of this test and made it
brittle.
Relax this coupling by only imposing a lower bound on the number of
state machine steps according to the test logic.

Note that this test could benefit from a more thorough refactoring, but
this particular change limits itself to what's necessary to ensure
resilience to an upcoming change.